backend/requirements.txt
Flask-SQLAlchemy
Flask-Migrate
Flask-Admin[sqlalchemy]
connexion[flask, swagger-ui, uvicorn]
python-dotenv
psycopg2-binary